+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -145,9 +145,12 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     if (name == null) {
       return;
     }
+    if (name == '/') {
+      name = 'root ("/")';
+    }
     _transaction = _hub.startTransaction(
       name,
-      'ui.load',
+      'navigation',
       waitForChildren: true,
       autoFinishAfter: Duration(seconds: 3),
     );