+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Unreleased
 
 * Feat: Improve configuration options of `SentryNavigatorObserver` (#684)
+* Feat: Add support for [Dio](https://pub.dev/packages/dio) (#688)
+* Bump: Sentry-Android to 5.5.2 and Sentry-Cocoa to 7.8.0 (#696)
+
+# 6.3.0-beta.1
+
+* Enha: Replace flutter default root name '/' with 'root' (#678)
+* Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
+* Fix: Use correct data/extras type in tracer (#693)
+* Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#692)
 
 # 6.3.0-alpha.1